backend/gRPC/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
bcrypt
pyjwt